revert some gen changes
