+++ Pipeline/utils.py
@@ -312,3 +312,13 @@ def profile_for_rgb(profile: rasterio.profiles.Profile) -> rasterio.profiles.Pro
     profile['blockysize'] = 256
     profile['nodata'] = 0
     return profile
+
+
+def supported_granule_type(_type: str) -> bool:
+    return _type in ["L1C", "L2A"]
+
+
+def get_metadata_path(granule_path: str, granule_type: str) -> str:
+    if supported_granule_type(granule_type):
+        raise NotImplemented
+    return granule_path + os.path.sep + ("MTD_MSIL2A.xml" if granule_type == "L2A" else "MTD_MSIL1C.xml")
